+++ python/src/iceberg/utils/schema_conversion.py
@@ -23,7 +23,7 @@ import logging
 from typing import Any
 
 from iceberg.schema import Schema
-from iceberg.types import (
+from iceberg.type import (
     BinaryType,
     BooleanType,
     DateType,
