+++ python/tests/test_transforms.py
@@ -23,7 +23,7 @@ import mmh3 as mmh3
 import pytest
 
 from iceberg import transforms
-from iceberg.types import (
+from iceberg.type import (
     BinaryType,
     BooleanType,
     DateType,
